requirements.txt
streamlit
email
pdfplumber
pytesseract
Pillow
docx
json
python-dotenv
